Update packages in CP
