+++ platforms/intellij/compose-playground/src/jvmMain/kotlin/foundry/intellij/compose/playground/MarkdownPlayground.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package slack.tooling.compose.playground
+package foundry.intellij.compose.playground
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
@@ -29,8 +29,7 @@ import org.jetbrains.jewel.foundation.theme.JewelTheme
 import org.jetbrains.jewel.intui.standalone.theme.IntUiTheme
 import org.jetbrains.jewel.ui.component.DefaultButton
 import org.jetbrains.jewel.ui.component.Text
-import slack.tooling.projectgen.ProjectGenUi
-import slack.tooling.projectgen.ProjectGenUi.ProjectGenApp
+import slack.tooling.markdown.ui.MarkdownContent
 
 fun main() = singleWindowApplication {
   var isDark by remember { mutableStateOf(false) }
@@ -39,19 +38,64 @@ fun main() = singleWindowApplication {
       DefaultButton(modifier = Modifier.padding(16.dp), onClick = { isDark = !isDark }) {
         Text("Toggle dark mode")
       }
-      ProjectGenApp(
-        rootDir = "rootDir",
-        events =
-          object : ProjectGenUi.Events {
-            override fun doOKAction() {
-              println("doOKAction")
-            }
-
-            override fun dismissDialogAndSync() {
-              println("dismissDialogAndSync")
-            }
-          },
-      )
+      MarkdownContent { MARKDOWN }
     }
   }
 }
+
+private const val MARKDOWN =
+  """
+  # Markdown Playground
+
+  ---
+
+  # This is an H1
+
+  ## This is an H2
+
+  ### This is an H3
+
+  #### This is an H4
+
+  ##### This is an H5
+
+  ###### This is an H6
+
+  This is a paragraph with some *italic* and **bold** text.
+
+  This is a paragraph with some `inline code`.
+
+  This is a paragraph with a [link](https://www.jetbrains.com/).
+
+  This is a code block:
+  ```kotlin
+  fun main() {
+    println("Hello, world!")
+  }
+  ```
+
+  > This is a block quote.
+
+  This is a divider
+
+  ---
+
+  The above was supposed to be a divider.
+
+  ~~This is strikethrough~~
+
+  This is an ordered list:
+  1. Item 1
+  2. Item 2
+  3. Item 3
+
+  This is an unordered list with dashes:
+  - Item 1
+  - Item 2
+  - Item 3
+
+  This is an unordered list with asterisks:
+  * Item 1
+  * Item 2
+  * Item 3
+"""
